# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: PortUrl API
  description: API for managing applications and categories in PortUrl
  version: 1.0.0
servers:
  - url: /
...
paths:
...
  /api/admin/import:
    post:
      tags:
        - Admin
      summary: Import apps and categories from YAML
      operationId: importData
      requestBody:
        required: true
        content:
          application/x-yaml:
            schema:
              $ref: '#/components/schemas/ExportData'
      responses:
        '204':
          description: No Content
  /api/admin/realms/{realm}/clients:
    get:
      tags:
        - Admin
      summary: List available clients in a target realm (auto-discovery)
      operationId: scanRealmClients
      parameters:
        - name: realm
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/KeycloakClientDto'
...
components:
  schemas:
    KeycloakClientDto:
      type: object
      properties:
        id:
          type: string
          description: Internal Keycloak UUID
        clientId:
          type: string
          description: Public client ID (e.g. "grafana")
        name:
          type: string
          description: Human readable name
...
    ExportData:
      type: object
      properties:
        categories:
          type: array
          items:
            $ref: '#/components/schemas/CategoryExport'
        applications:
          type: array
          items:
            $ref: '#/components/schemas/ApplicationExport'
    CategoryExport:
      type: object
      properties:
        name:
          type: string
        sortOrder:
          type: integer
        applicationSortMode:
          type: string
          enum: [CUSTOM, ALPHABETICAL]
        description:
          type: string
    ApplicationExport:
      type: object
      properties:
        name:
          type: string
        url:
          type: string
        categories:
          type: array
          items:
            type: string
        roles:
          type: array
          items:
            type: string
        icon:
          type: string
          description: Base64 encoded icon content.
        clientId:
          type: string
          nullable: true
        realm:
          type: string
          nullable: true